ai-flight-tracker/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
pytest>=7.4.0
pytest-flask>=1.3.0
